OCPBUGS-64775: use CAPZ to provision ssh rule
A change to CAPZ[0], creates an SSH rule if one is not specified in
the cluster spec. Prior to this commit, we had been creating the
SSH rule with installer SDK hooks, which is still somewhat necessary
to add the inbound NAT rules, because we are not yet using CAPZ
to provision a public load balancer.

But we can use CAPZ to just create the rule, which will stop CAPZ
from preventing a redundant SSH rule which we were leaking during
bootstrap destroy.

This change will also result in creating an SSH rule for private clusters
which is fine, and something we do on other providers.

0: https://github.com/kubernetes-sigs/cluster-api-provider-azure/pull/5525